move balance commands to separate folder
