+++ src/main/java/frc/robot/commands/Balance/BalancePIDCommand.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.Balance;
 
 import com.revrobotics.CANSparkMax.IdleMode;
